Refactored card-utils into some other files.
