+++ src/card/lib/card-utils.js
@@ -1,77 +1,20 @@
 /* @flow */
 
-import { noop, values } from '@krakenjs/belter';
-import cardValidator from 'card-validator';
+import { values } from '@krakenjs/belter';
 
-import type { CardType, CardNavigation, InputState, FieldValidity, InputEvent, Card, ExtraFields } from '../types';
+import type { InputState, FieldValidity, Card, ExtraFields } from '../types';
 import {
     CARD_ERRORS,
     CARD_FIELD_TYPE,
-    DEFAULT_CARD_TYPE,
     DEFAULT_STYLE,
     FIELD_STYLE,
     FILTER_CSS_SELECTORS,
     FILTER_CSS_VALUES,
     GQL_ERRORS,
     VALID_EXTRA_FIELDS,
-    VALIDATOR_TO_TYPE_MAP,
     ALLOWED_ATTRIBUTES
 } from '../constants';
 import { getLogger } from '../../lib';
-
-// Add additional supported card types
-cardValidator.creditCardType.addCard({
-    code: {
-        name: 'CVV',
-        size: 3
-    },
-    gaps:     [ 4, 8, 12 ],
-    lengths:  [ 16, 18, 19 ],
-    niceType: 'Carte Bancaire',
-    patterns: [],
-    type:     'cb-nationale'
-});
-
-cardValidator.creditCardType.addCard({
-    code: {
-        name: 'CVV',
-        size: 3
-    },
-    gaps:     [ 4, 8, 12, 16 ],
-    lengths:  [ 19 ],
-    niceType: 'Carte Aurore',
-    patterns: [],
-    type:     'cetelem'
-});
-
-cardValidator.creditCardType.addCard({
-    code: {
-        name: '',
-        size: 0
-    },
-    gaps:     [ 4, 8, 12, 16 ],
-    lengths:  [ 17 ],
-    niceType: 'Cofinoga ou Privilège',
-    patterns: [],
-    type:     'cofinoga'
-});
-
-cardValidator.creditCardType.addCard({
-    code: {
-        name: '',
-        size: 0
-    },
-    gaps:     [ 4, 8 ],
-    lengths:  [ 8, 9 ],
-    niceType: '4 étoiles',
-    patterns: [],
-    type:     'cofidis'
-});
-
-export const defaultNavigation : CardNavigation = {
-    next:     () => noop,
-    previous: () => noop
-};
 
 export const defaultInputState : InputState = {
     inputValue:         '',
@@ -89,7 +32,6 @@ export function splice(str : string, idx : number, insert : string) : string {
     return str.slice(0, idx) + insert + str.slice(idx);
 }
 
-
 export function assertType(assertion : () => void, errorMsg : string) : void | TypeError {
     if (!assertion) {
         throw new TypeError(errorMsg);
@@ -103,39 +45,6 @@ export function assertString<T>(...args : T) : void | TypeError {
 
 export function removeSpaces(value : string) : string {
     return value.replace(/\s/g, '');
-}
-
-// Detect the card type metadata for a card number
-export function detectCardType(number : string) : CardType {
-    if (number.length > 0) {
-        const cardTypes = cardValidator.creditCardType.default(number);
-        if (cardTypes.length > 0) {
-            return cardTypes[0];
-        }
-    }
-    return DEFAULT_CARD_TYPE;
-}
-
-// Mask a card number for display given a card type. If a card type is
-// not provided, attempt to detect it and mask based on that type.
-export function maskCard(number : string, cardType? : CardType) : string {
-    assertString(number);
-    number = number.trim().replace(/[^0-9]/g, '').replace(/\s/g, '');
-    // $FlowFixMe
-    const gaps = cardType?.gaps || detectCardType(number)?.gaps;
-
-    if (gaps) {
-        for (let idx = 0; idx < gaps.length; idx++) {
-            const splicePoint = gaps[idx] + idx;
-            if (splicePoint > number.length - 1) {
-                // We're beyond the end of the number
-                break;
-            }
-
-            number = splice(number, splicePoint, ' ');
-        }
-    }
-    return number;
 }
 
 // Return the last 4 digits of a valid card number
@@ -196,6 +105,15 @@ function isValidValue(value : string | number) : boolean {
 // from https://github.com/braintree/inject-stylesheet/blob/main/src/lib/validate-selector.ts
 function isValidSelector(selector : string) : boolean {
     return !FILTER_CSS_SELECTORS.some((regex) => regex.test(selector));
+}
+
+export function isValidAttribute(attribute: string) : boolean {
+    if(!ALLOWED_ATTRIBUTES.includes(attribute.toLocaleLowerCase())) {
+        getLogger().warn('attribute_warning', { warn: `HTML Attribute "${attribute}" was ignored. See allowed attribute list.`})
+        return false
+    };
+
+    return true
 }
 
 export function filterStyle(style : Object) : Object {
@@ -268,84 +186,6 @@ export function checkForNonDigits(value : string) : boolean {
     return (/\D/g).test(removeSpaces(value));
 }
 
-export function getCvvLength(cardType? : CardType) : number {
-    if (cardType && typeof cardType === 'object') {
-        const { code } = cardType;
-
-        if (typeof code === 'object') {
-            const { size } = code;
-
-            if (typeof size === 'number') {
-                return size;
-            }
-        }
-    }
-
-    return 3;
-}
-
-export function checkCardEligibility(value : string, cardType : CardType) : boolean  {
-    // check if the card type is eligible
-    const fundingEligibility = window.xprops.fundingEligibility;
-    const type = VALIDATOR_TO_TYPE_MAP[cardType.type];
-    // only mark as ineligible if the card vendor is explicitly set to not be eligible
-    if (type && fundingEligibility?.card?.eligible) {
-        const vendor = fundingEligibility.card.vendors?.[type];
-        if (vendor && !vendor.eligible) {
-            return false;
-        }
-    }
-    // otherwise default to be eligible
-    return true;
-}
-
-export function checkCardNumber(value : string) : {| isValid : boolean, isPotentiallyValid : boolean |} {
-    const { number } = cardValidator;
-
-    const {isValid, isPotentiallyValid} = number(value);
-
-    return {
-        isValid,
-        isPotentiallyValid
-    }
-}
-
-export function checkCVV(value : string, cardType : CardType) : {| isValid : boolean, isPotentiallyValid : boolean |} {
-    let isValid = false;
-    if (value.length === getCvvLength(cardType)) {
-        isValid = true;
-    }
-    return {
-        isValid,
-        isPotentiallyValid: true
-    };
-}
-
-export function checkName(value : string) : {| isValid : boolean, isPotentiallyValid : boolean |} {
-    const { cardholderName } = cardValidator
-
-    return cardholderName(value)
-}
-
-export function checkExpiry(value : string) : {| isValid : boolean, isPotentiallyValid : boolean |} {
-    const { expirationDate } = cardValidator;
-    const { isValid } = expirationDate(value);
-
-    return {
-        isValid,
-        isPotentiallyValid: true
-    };
-}
-
-export function checkPostalCode(value : string, minLength? : number) : {| isValid : boolean, isPotentiallyValid : boolean |} {
-    const { postalCode } = cardValidator;
-    const { isValid } = postalCode(value, {minLength})
-    return {
-        isValid,
-        isPotentiallyValid: true
-    };
-}
-
 export function setErrors({ isCardEligible, isNumberValid, isCvvValid, isExpiryValid, isNameValid, isPostalCodeValid, gqlErrorsObject = {} } : {| isCardEligible? : boolean, isNumberValid? : boolean, isCvvValid? : boolean, isExpiryValid? : boolean, isNameValid? : boolean, isPostalCodeValid? : boolean, gqlErrorsObject? : {| field : string, errors : [] |} |}) : [$Values<typeof CARD_ERRORS>] | [] {
     const errors = [];
     const { field, errors: gqlErrors } = gqlErrorsObject;
@@ -401,48 +241,6 @@ export function setErrors({ isCardEligible, isNumberValid, isCvvValid, isExpiryV
     return errors;
 }
 
-// Move cursor within a field
-export function moveCursor(element : HTMLInputElement, start : number, end? : number) : void {
-    window.requestAnimationFrame(() => {
-        element.selectionStart = start;
-        element.selectionEnd = end ?? start;
-    });
-}
-
-// Navigation helper to go to the next field putting the cursor at the start
-export function goToNextField(ref : {| current : {| base : HTMLInputElement |} |}) : () => void {
-    return () => {
-        moveCursor(ref.current.base, 0);
-        setTimeout(() => ref.current.base.focus());
-    };
-}
-
-// Navigation helper to go to the previous field putting the curser at the end
-export function goToPreviousField(ref : {| current : {| base : HTMLInputElement |} |}) : () => void {
-    return () => {
-        const { value } = ref.current.base;
-
-        if (value) {
-            const valueLength = value.length;
-            moveCursor(ref.current.base, valueLength);
-        }
-        setTimeout(() => ref.current.base.focus());
-    };
-}
-
-// Navigate between fields using the arrow keys and/or the backspace
-export function navigateOnKeyDown(event : InputEvent, navigation : CardNavigation) : void {
-    const { target: { value, selectionStart, selectionEnd }, key } = event;
-
-    if (selectionStart === 0 && (value.length === 0 || value.length !== selectionEnd)  && [ 'Backspace', 'ArrowLeft' ].includes(key)) {
-        navigation.previous();
-    }
-
-    if (selectionStart === value.length && [ 'ArrowRight' ].includes(key)) {
-        navigation.next();
-    }
-}
-
 // Format expity date to MM/YYYY
 export function convertDateFormat(date : string) : string {
     const trimmedDate = removeSpaces(date);
@@ -455,63 +253,6 @@ export function convertDateFormat(date : string) : string {
     }
 
     return formattedDate;
-}
-
-
-// Safari (both iOS and Desktop) has an unconvential behavior,
-// where it won't let an iframe that includes an input get
-// focus programatically from outisde of the input.
-// Big props to the devs at Stripe that figured out
-// you run this selection range hack to force the focus back
-// onto the input.
-function applyFocusWorkaroundForSafari (input : HTMLInputElement) {
-    const inputIsEmptyInitially = input.value === '';
-
-    // Safari can't set selection if the input is empty
-    if (inputIsEmptyInitially) {
-        input.value = ' ';
-    }
-
-    const start = input.selectionStart;
-    const end = input.selectionEnd;
-
-    input.setSelectionRange(0, 0);
-    input.setSelectionRange(start, end);
-
-    if (inputIsEmptyInitially) {
-        input.value = '';
-    }
-}
-
-export function autoFocusOnFirstInput(input? : HTMLInputElement) {
-    if (!input) {
-        return;
-    }
-
-    let timeoutID = null;
-
-    window.addEventListener('focus', () => {
-        // the set timeout is required here, because in some browsers (Firefox, for instance)
-        // when tabbing backward into the iframe, it will have the html element focussed
-        // initially, but then passes focus to the input
-        timeoutID = setTimeout(() => {
-            timeoutID = null;
-
-            applyFocusWorkaroundForSafari(input);
-
-            // for Safari, setting the selection range is enough to give
-            // it focus, but Firefox requires an explicit focus call.
-            // Also, just calling `focus` on Safari does not work at all
-            input.focus();
-        }, 1);
-    });
-
-    window.addEventListener('focusin', (event) => {
-        if (timeoutID && event.target instanceof HTMLInputElement) {
-            clearTimeout(timeoutID);
-            timeoutID = null;
-        }
-    });
 }
 
 // Function that returns the field value in the correct format
@@ -585,13 +326,4 @@ export function filterExtraFields(extraData : Object) : ExtraFields | Object {
         }
         return acc;
     }, {});
-}
-
-export function isValidAttribute(attribute: string) : boolean {
-    if(!ALLOWED_ATTRIBUTES.includes(attribute.toLocaleLowerCase())) {
-        getLogger().warn('attribute_warning', { warn: `HTML Attribute "${attribute}" was ignored. See allowed attribute list.`})
-        return false
-    };
-
-    return true
 }
