+++ src/card/lib/index.js
@@ -1,5 +1,7 @@
 /* @flow */
 
+export * from './card-checks';
+export * from './card-focus';
+export * from './card-utils';
 export * from './exports';
 export * from './methods';
-export * from './card-utils';
